JsonModel 분리
